requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.2.25
django-cors-headers==3.10.1
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
importlib-resources==5.4.0
PyJWT==1.7.1
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.1.1
zipp==3.6.0
